before merge

+++ miniRT/incs/miniRT.h
@@ -71,7 +71,7 @@ void		init_camera_ray(t_ray *cam_ray, t_scene *scene);
 
 /************************************************** COLOR **********************************************************/
 
-int			get_color_pixel(t_obj *obj, t_scene *scene, t_ray *ray, unsigned int *color, double pixel_shadow, int rebound);
+int			get_color_pixel(t_obj *obj, t_scene *scene, t_ray *ray, long unsigned int *color, double pixel_shadow, int rebound);
 void		draw_pixel(t_image *image, int x, int y, int color);
 int			create_trgb(int transparency, int red, int green, int blue);
 void		check_limit_color(t_rgb *color);
